feat: fix share houses edit page
